refactor(auth): rename sso_clients.py to oidc_client.py (OIDC-specific)
